+++ docs/release_notes.md
@@ -1,4 +1,7 @@
 # Release Notes
+## v1.0.22
+**ID fields**: Every datum by default has an `id` field attached. You no longer have to specify an `id` field when mapping data (by you cannot use the field `id`).
+**Bug fixes**: Numerous bugs are now squashed involving error handling.
 
 ## v1.0.14
 **Progressive Maps**: Maps can now be built progressively in Atlas. See the progressive map documentation for more information.